Prevent video iframe from adding to browser history
When an iframe's src changes, the browser seems to add its previous src
to the history.

As a result, if you play a couple of videos on a page you then have to
click "back" a number of times in order to get to the previous page.

To fix that, let's create a completely new iframe when showing a video.