PLANNER-2637: Upgrade to Antora 3
Need to upgrade Node version since Javascript introduced
new syntax the old version could not recognized. Because
https://github.com/eirslett/frontend-maven-plugin/issues/1005
was resolved, I was able to remove the exec plugin and replace
it with a npx execution in the frontend-maven-plugin. Changed
the ui-bundle to the Antora 3 ui-bundle. Updated supplemental-ui
to match new expected layout.

Because Lunr extension is officially supported by Antora, its license
is changed from MIT to MPL-2.0 (which matches the license the
supplemental-ui and the ui-bundle zip are under).